Use explicit instantation to move fmt lib to cpp files.
