+++ CMakeLists.txt
@@ -66,19 +66,28 @@ set(SOURCE_FILES
     src/log_surgeon/FileReader.hpp
     src/log_surgeon/finite_automata/Capture.hpp
     src/log_surgeon/finite_automata/DeterminizationConfiguration.hpp
+    src/log_surgeon/finite_automata/Dfa.cpp
     src/log_surgeon/finite_automata/Dfa.hpp
+    src/log_surgeon/finite_automata/DfaState.cpp
     src/log_surgeon/finite_automata/DfaState.hpp
     src/log_surgeon/finite_automata/DfaStatePair.hpp
+    src/log_surgeon/finite_automata/DfaTransition.cpp
     src/log_surgeon/finite_automata/DfaTransition.hpp
+    src/log_surgeon/finite_automata/Nfa.cpp
     src/log_surgeon/finite_automata/Nfa.hpp
+    src/log_surgeon/finite_automata/NfaSpontaneousTransition.cpp
     src/log_surgeon/finite_automata/NfaSpontaneousTransition.hpp
+    src/log_surgeon/finite_automata/NfaState.cpp
     src/log_surgeon/finite_automata/NfaState.hpp
     src/log_surgeon/finite_automata/PrefixTree.cpp
     src/log_surgeon/finite_automata/PrefixTree.hpp
+    src/log_surgeon/finite_automata/RegexAST.cpp
     src/log_surgeon/finite_automata/RegexAST.hpp
     src/log_surgeon/finite_automata/RegisterHandler.hpp
+    src/log_surgeon/finite_automata/RegisterOperation.cpp
     src/log_surgeon/finite_automata/RegisterOperation.hpp
     src/log_surgeon/finite_automata/StateType.hpp
+    src/log_surgeon/finite_automata/TagOperation.cpp
     src/log_surgeon/finite_automata/TagOperation.hpp
     src/log_surgeon/finite_automata/UnicodeIntervalTree.hpp
     src/log_surgeon/finite_automata/UnicodeIntervalTree.tpp
