+++ src/log_surgeon/finite_automata/DfaTransition.hpp
@@ -11,9 +11,6 @@
 #include <log_surgeon/finite_automata/RegisterOperation.hpp>
 #include <log_surgeon/finite_automata/StateType.hpp>
 
-#include <fmt/core.h>
-#include <fmt/format.h>
-
 namespace log_surgeon::finite_automata {
 template <StateType state_type>
 class DfaState;
@@ -54,26 +51,6 @@ private:
     std::vector<RegisterOperation> m_reg_ops;
     DfaState<state_type> const* m_dest_state{nullptr};
 };
-
-template <StateType state_type>
-auto DfaTransition<state_type>::serialize(
-        std::unordered_map<DfaState<state_type> const*, uint32_t> const& state_ids
-) const -> std::optional<std::string> {
-    if (false == state_ids.contains(m_dest_state)) {
-        return std::nullopt;
-    }
-
-    std::vector<std::string> transformed_ops;
-    for (auto const& reg_op : m_reg_ops) {
-        auto const optional_serialized_op{reg_op.serialize()};
-        if (false == optional_serialized_op.has_value()) {
-            return std::nullopt;
-        }
-        transformed_ops.emplace_back(optional_serialized_op.value());
-    }
-
-    return fmt::format("-({})->{}", fmt::join(transformed_ops, ","), state_ids.at(m_dest_state));
-}
 }  // namespace log_surgeon::finite_automata
 
 #endif  // LOG_SURGEON_FINITE_AUTOMATA_DFATRANSITION_HPP
