+++ src/log_surgeon/finite_automata/Nfa.cpp
@@ -1,0 +1,30 @@
+#include <log_surgeon/finite_automata/Nfa.hpp>
+#include <log_surgeon/finite_automata/NfaState.hpp>
+
+#include <fmt/core.h>
+#include <fmt/format.h>
+
+namespace log_surgeon::finite_automata {
+template <typename TypedNfaState>
+auto Nfa<TypedNfaState>::serialize() const -> std::optional<std::string> {
+    auto const traversal_order = get_bfs_traversal_order();
+
+    std::unordered_map<TypedNfaState const*, uint32_t> state_ids;
+    for (auto const* state : traversal_order) {
+        state_ids.emplace(state, state_ids.size());
+    }
+
+    std::vector<std::string> serialized_states;
+    for (auto const* state : traversal_order) {
+        auto const optional_serialized_state{state->serialize(state_ids)};
+        if (false == optional_serialized_state.has_value()) {
+            return std::nullopt;
+        }
+        serialized_states.emplace_back(optional_serialized_state.value());
+    }
+    return fmt::format("{}\n", fmt::join(serialized_states, "\n"));
+}
+
+template auto Nfa<ByteNfaState>::serialize() const -> std::optional<std::string>;
+template auto Nfa<Utf8NfaState>::serialize() const -> std::optional<std::string>;
+}  // namespace log_surgeon::finite_automata
