+++ src/log_surgeon/finite_automata/RegexAST.hpp
@@ -19,9 +19,6 @@
 #include <log_surgeon/finite_automata/TagOperation.hpp>
 #include <log_surgeon/finite_automata/UnicodeIntervalTree.hpp>
 
-#include <fmt/core.h>
-#include <fmt/ranges.h>
-#include <fmt/xchar.h>
 #include <gsl/pointers>
 
 namespace log_surgeon::finite_automata {
@@ -157,25 +154,7 @@ protected:
     RegexAST(RegexAST&& rhs) noexcept = delete;
     auto operator=(RegexAST&& rhs) noexcept -> RegexAST& = delete;
 
-    [[nodiscard]] auto serialize_negative_captures() const -> std::u32string {
-        if (m_negative_captures.empty()) {
-            return U"";
-        }
-
-        auto const transformed_negative_captures{
-                m_negative_captures | std::ranges::views::transform([](Capture const* capture) {
-                    return fmt::format("<~{}>", capture->get_name());
-                })
-        };
-        auto const negative_captures_string{
-                fmt::format("{}", fmt::join(transformed_negative_captures, ""))
-        };
-
-        return fmt::format(
-                U"{}",
-                std::u32string(negative_captures_string.begin(), negative_captures_string.end())
-        );
-    }
+    [[nodiscard]] auto serialize_negative_captures() const -> std::u32string;
 
 private:
     std::vector<Capture const*> m_subtree_positive_captures;
@@ -737,11 +716,6 @@ private:
 };
 
 template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTEmpty<TypedNfaState>::serialize() const -> std::u32string {
-    return fmt::format(U"{}", RegexAST<TypedNfaState>::serialize_negative_captures());
-}
-
-template <typename TypedNfaState>
 RegexASTLiteral<TypedNfaState>::RegexASTLiteral(uint32_t character) : m_character(character) {}
 
 template <typename TypedNfaState>
@@ -751,15 +725,6 @@ void RegexASTLiteral<TypedNfaState>::add_to_nfa(
         [[maybe_unused]] bool const descendent_of_repetition
 ) const {
     nfa->add_root_interval(Interval(m_character, m_character), end_state);
-}
-
-template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTLiteral<TypedNfaState>::serialize() const -> std::u32string {
-    return fmt::format(
-            U"{}{}",
-            static_cast<char32_t>(m_character),
-            RegexAST<TypedNfaState>::serialize_negative_captures()
-    );
 }
 
 template <typename TypedNfaState>
@@ -782,16 +747,6 @@ void RegexASTInteger<TypedNfaState>::add_to_nfa(
         [[maybe_unused]] bool const descendent_of_repetition
 ) const {
     throw std::runtime_error("Unsupported");
-}
-
-template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTInteger<TypedNfaState>::serialize() const -> std::u32string {
-    auto const digits_string = fmt::format("{}", fmt::join(m_digits, ""));
-    return fmt::format(
-            U"{}{}",
-            std::u32string(digits_string.begin(), digits_string.end()),
-            RegexAST<TypedNfaState>::serialize_negative_captures()
-    );
 }
 
 template <typename TypedNfaState>
@@ -820,16 +775,6 @@ void RegexASTOr<TypedNfaState>::add_to_nfa(
 }
 
 template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTOr<TypedNfaState>::serialize() const -> std::u32string {
-    return fmt::format(
-            U"({})|({}){}",
-            nullptr != m_left ? m_left->serialize() : U"null",
-            nullptr != m_right ? m_right->serialize() : U"null",
-            RegexAST<TypedNfaState>::serialize_negative_captures()
-    );
-}
-
-template <typename TypedNfaState>
 RegexASTCat<TypedNfaState>::RegexASTCat(
         std::unique_ptr<RegexAST<TypedNfaState>> left,
         std::unique_ptr<RegexAST<TypedNfaState>> right
@@ -854,16 +799,6 @@ void RegexASTCat<TypedNfaState>::add_to_nfa(
     nfa->set_root(intermediate_state);
     m_right->add_to_nfa_with_negative_captures(nfa, end_state, descendent_of_repetition);
     nfa->set_root(saved_root);
-}
-
-template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTCat<TypedNfaState>::serialize() const -> std::u32string {
-    return fmt::format(
-            U"{}{}{}",
-            nullptr != m_left ? m_left->serialize() : U"null",
-            nullptr != m_right ? m_right->serialize() : U"null",
-            RegexAST<TypedNfaState>::serialize_negative_captures()
-    );
 }
 
 template <typename TypedNfaState>
@@ -918,20 +853,6 @@ void RegexASTMultiplication<TypedNfaState>::add_to_nfa(
 }
 
 template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTMultiplication<TypedNfaState>::serialize() const -> std::u32string {
-    auto const min_string = std::to_string(m_min);
-    auto const max_string = std::to_string(m_max);
-
-    return fmt::format(
-            U"({}){{{},{}}}{}",
-            nullptr != m_operand ? m_operand->serialize() : U"null",
-            std::u32string(min_string.begin(), min_string.end()),
-            is_infinite() ? U"inf" : std::u32string(max_string.begin(), max_string.end()),
-            RegexAST<TypedNfaState>::serialize_negative_captures()
-    );
-}
-
-template <typename TypedNfaState>
 auto RegexASTCapture<TypedNfaState>::add_to_nfa(
         Nfa<TypedNfaState>* nfa,
         TypedNfaState* dest_state,
@@ -980,19 +901,6 @@ auto RegexASTCapture<TypedNfaState>::add_to_nfa(
     m_capture_regex_ast
             ->add_to_nfa_with_negative_captures(nfa, capture_end_state, descendent_of_repetition);
     nfa->set_root(initial_root);
-}
-
-template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTCapture<TypedNfaState>::serialize() const -> std::u32string {
-    auto const capture_name_u32{
-            std::u32string(m_capture->get_name().cbegin(), m_capture->get_name().cend())
-    };
-    return fmt::format(
-            U"({})<{}>{}",
-            m_capture_regex_ast->serialize(),
-            capture_name_u32,
-            RegexAST<TypedNfaState>::serialize_negative_captures()
-    );
 }
 
 template <typename TypedNfaState>
@@ -1125,37 +1033,6 @@ void RegexASTGroup<TypedNfaState>::add_to_nfa(
     for (auto const& [begin, end] : merged_ranges) {
         nfa->get_root()->add_interval(Interval(begin, end), end_state);
     }
-}
-
-template <typename TypedNfaState>
-[[nodiscard]] auto RegexASTGroup<TypedNfaState>::serialize() const -> std::u32string {
-    std::u32string ranges_serialized;
-    if (m_is_wildcard) {
-        ranges_serialized += U"*";
-    } else {
-        auto const transformed_ranges
-                = m_ranges
-                  | std::ranges::views::transform([](std::pair<uint32_t, uint32_t> const& range) {
-                        auto const [begin, end] = range;
-                        return fmt::format(
-                                U"{}-{}",
-                                static_cast<char32_t>(begin),
-                                static_cast<char32_t>(end)
-                        );
-                    });
-        for (auto const& range_u32string : transformed_ranges) {
-            if (false == ranges_serialized.empty()) {
-                ranges_serialized += U", ";  // Add separator
-            }
-            ranges_serialized += range_u32string;
-        }
-    }
-    return fmt::format(
-            U"[{}{}]{}",
-            m_negate ? U"^" : U"",
-            ranges_serialized,
-            RegexAST<TypedNfaState>::serialize_negative_captures()
-    );
 }
 }  // namespace log_surgeon::finite_automata
